// src/app/Components/Three/components/Island/types.ts
import { GLTF } from "three-stdlib";

export type GLTFResult = GLTF & {
  nodes: {
    base001: THREE.Mesh;
    grass001: THREE.Mesh;
    path004: THREE.Mesh;
    Cylinder001: THREE.Mesh;
    Cylinder001_1: THREE.Mesh;
    tree005: THREE.Mesh;
    tree006: THREE.Mesh;
    rock046: THREE.Mesh;
    rock047: THREE.Mesh;
    path005: THREE.Mesh;
    rock048: THREE.Mesh;
    rock049: THREE.Mesh;
    rock050: THREE.Mesh;
    rock051: THREE.Mesh;
    rock052: THREE.Mesh;
    rock053: THREE.Mesh;
    rock054: THREE.Mesh;
    rock055: THREE.Mesh;
    rock056: THREE.Mesh;
    rock057: THREE.Mesh;
    rock058: THREE.Mesh;
    rock059: THREE.Mesh;
    rock060: THREE.Mesh;
    rock061: THREE.Mesh;
    bush011: THREE.Mesh;
    bush012: THREE.Mesh;
    bush013: THREE.Mesh;
    bush014: THREE.Mesh;
    bush015: THREE.Mesh;
    bench001: THREE.Mesh;
    bush016: THREE.Mesh;
    bush017: THREE.Mesh;
    bush018: THREE.Mesh;
    path006: THREE.Mesh;
    rock062: THREE.Mesh;
    rock063: THREE.Mesh;
    bush019: THREE.Mesh;
    bush020: THREE.Mesh;
    rock064: THREE.Mesh;
    rock065: THREE.Mesh;
    rock066: THREE.Mesh;
    rock067: THREE.Mesh;
    rock068: THREE.Mesh;
    rock069: THREE.Mesh;
    rock070: THREE.Mesh;
    rock071: THREE.Mesh;
    rock072: THREE.Mesh;
    rock073: THREE.Mesh;
    rock075: THREE.Mesh;
    rock076: THREE.Mesh;
    rock077: THREE.Mesh;
    rock078: THREE.Mesh;
    rock079: THREE.Mesh;
    Mesh004: THREE.Mesh;
    Mesh004_1: THREE.Mesh;
    Mesh005: THREE.Mesh;
    Mesh005_1: THREE.Mesh;
    Mesh006: THREE.Mesh;
    Mesh006_1: THREE.Mesh;
    Plane023: THREE.Mesh;
  };
  materials: {
    ["Color texture"]: THREE.MeshStandardMaterial;
    ["Color texture"]: THREE.MeshStandardMaterial;
    light: THREE.MeshStandardMaterial;
    trunk: THREE.MeshStandardMaterial;
  };
};